refactor: data 변경 및 숙소 조건별 검색 쿼리 수정

+++ backend/src/main/java/team15/airbnb/accommodation/application/AccommodationService.java
@@ -39,8 +39,9 @@ public class AccommodationService {
 
     public SearchAccommodationsByOptionsResponse searchByOptions(SearchAccommodationsOptionsRequest request, Long userId) {
         List<AccommodationSimpleInfoResponse> searchResult = accommodationRepository.findByOptions(request, userId);
-        Long totalCount = accommodationRepository.getAccommodationCounts(request);
+        int totalCount = accommodationRepository.getAccommodationCounts(request);
         boolean hasNext = hasNext(searchResult.size(), request.getLimit());
+
         AccommodationListResponse accommodationListResponse = new AccommodationListResponse(totalCount, searchResult);
 
         if (hasNext) {
